Adicionar suporte para criação de GIFs e codificação de imagens em Base64
