+++ src/tsimg_gif.cpp
@@ -3,36 +3,21 @@
 
 #include <stb_image.h>
 #include <stb_image_resize2.h>
-#include "image_utils.h"
 #include "gif.h"
+#include "tsimg_gif.h"
 #include <iostream>
 #include <fstream>
 
-
-/**
- * @brief Creates a GIF from a sequence of images.
- *
- * This function takes a list of image file paths, resizes each image to the specified width and height,
- * and combines them into a single GIF file.
- *
- * @param output_filename The name of the output GIF file.
- * @param image_paths A vector of strings containing the file paths of the images to be included in the GIF.
- * @param width The width to which each image should be resized.
- * @param height The height to which each image should be resized.
- * @param debug If true, debug information will be printed to the standard output.
- * @return true if the GIF was created successfully, false otherwise.
- */
 bool createGif(const std::string& output_filename, const std::vector<std::string>& image_paths, bool debug) {
     if (image_paths.empty()) {
-        if (debug) std::cerr << "No images provided." << std::endl;
+        if (debug) std::cerr << "Error: No images provided." << std::endl;
         return false;
     }
 
     int width = 0, height = 0, channels = 0;
 
-    // Carregar a primeira imagem para obter as dimensões
     if (debug) std::cout << "Loading first image to get dimensions..." << std::endl;
-    unsigned char* first_image = stbi_load(image_paths[0].c_str(), &width, &height, &channels, 4); // Carregar com 4 canais (RGBA)
+    unsigned char* first_image = stbi_load(image_paths[0].c_str(), &width, &height, &channels, 4);
     if (!first_image) {
         if (debug) std::cerr << "Failed to load image: " << image_paths[0] << std::endl;
         return false;
@@ -73,14 +58,6 @@ bool createGif(const std::string& output_filename, const std::vector<std::string
     return true;
 }
 
-
-/**
- * @brief Codifica uma imagem em formato Base64.
- * 
- * @param imagePath Caminho para a imagem.
- * @param debug Ativa o modo de depuração para imprimir detalhes da operação.
- * @return A string codificada em Base64 representando a imagem.
- */
 std::string encodeImageToBase64(const std::string& imagePath, bool debug) {
     if (debug) {
         std::cout << "Encoding image to Base64: " << imagePath << std::endl;
@@ -93,6 +70,11 @@ std::string encodeImageToBase64(const std::string& imagePath, bool debug) {
     }
 
     std::vector<unsigned char> buffer((std::istreambuf_iterator<char>(file)), std::istreambuf_iterator<char>());
+
+    if (buffer.empty()) {
+        std::cerr << "Error: File is empty or could not be read: " << imagePath << std::endl;
+        return "";
+    }
 
     static const char base64Chars[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
     std::string base64;
@@ -109,7 +91,7 @@ std::string encodeImageToBase64(const std::string& imagePath, bool debug) {
     while (base64.size() % 4) base64.push_back('=');
 
     if (debug) {
-        std::cout << "Image encoded successfully." << std::endl;
+        std::cout << "Image encoded successfully: " << imagePath << std::endl;
     }
 
     return base64;
